Moved non-page components out of 'pages' folder as it caused an error in pre-rendering
