+++ src/components/Text.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react"
-import data from '../translation.json'
+import data from '../pages/translation.json'
 
 type Props = {
     text: string,
@@ -18,13 +18,13 @@ const Text: FC<Props> = (props) => {
 
     let userLanguage = props.lang
 
-    // if (userLanguage.includes('en')) {
-        // userLanguage = 'en'
-    // } else if (userLanguage.includes('pl')) {
-        // userLanguage = 'pl'
-    // } else {
+    if (userLanguage.includes('en')) {
         userLanguage = 'en'
-    // }
+    } else if (userLanguage.includes('pl')) {
+        userLanguage = 'pl'
+    } else {
+        userLanguage = 'en'
+    }
 
     const translation = translations[props.text][userLanguage];
